+++ core/src/init/InitialConditions_restart.cpp
@@ -235,7 +235,7 @@ public:
         std::cout << "Restart mesh : " << input_lmesh.getNumOctants() << " octs." << std::endl;
 
         // Refine until level_max using analytical markers
-        for (uint8_t level=level_min; level<level_max; ++level)
+        for (uint32_t level=level_min; level<level_max; ++level)
         {
             const LightOctree& lmesh = pmesh.getLightOctree();
             uint32_t nbOcts = lmesh.getNumOctants();
